+++ terrain/voxel_viewer.cpp
@@ -135,7 +135,11 @@ void VoxelViewer::_notification(int p_what) {
 	switch (p_what) {
 		case NOTIFICATION_ENTER_TREE: {
 			if (!Engine::get_singleton()->is_editor_hint() || _enabled_in_editor) {
-				_viewer_id = VoxelEngine::get_singleton().add_viewer();
+				if (!_pending_deferred_unregistration) {
+					_viewer_id = VoxelEngine::get_singleton().add_viewer();
+				} else {
+					ZN_ASSERT_RETURN(VoxelEngine::get_singleton().viewer_exists(_viewer_id));
+				}
 				sync_all_parameters();
 				// VoxelEngine::get_singleton().sync_viewers_task_priority_data();
 			}
@@ -143,11 +147,21 @@ void VoxelViewer::_notification(int p_what) {
 
 		case NOTIFICATION_EXIT_TREE:
 			if (!Engine::get_singleton()->is_editor_hint() || _enabled_in_editor) {
-				// TODO When users reparent nodes, adding/removing viewers causes some suboptimal situations.
-				// We could mitigate this use case by putting viewers into an inactive group, so they keep their ID, so
-				// when reparenting happens, they will flip on and off. From the perspective of terrain's viewer pairing
-				// logic, it will be as if nothing special happened and it won't cause unnecessary reload/re-refcount.
-				VoxelEngine::get_singleton().remove_viewer(_viewer_id);
+				if (!_pending_deferred_unregistration) {
+					// When users reparent nodes, Godot triggers an EXIT_TREE, followed by a separate ENTER_TREE. So
+					// reparenting viewers is indistinguishable from solely adding, or removing a node. If we unregister
+					// right now, it leads to unexpected issues with terrain pairing, like the whole area reloading
+					// because the viewer got unregistered and re-registered. Since reparenting usually takes place with
+					// `remove_child` immediately followed by `add_child`, we can workaround the issue by deferring
+					// unregistration. Unfortunately, we then have to workaround the case the node gets deleted
+					// (`remove_child` followed by `free`, or `queue_free`), otherwise the deferred call could end
+					// badly.
+					_pending_deferred_unregistration = true;
+					callable_mp_static(&VoxelViewer::unregister_deferred_callback)
+							.bind(get_instance_id(), Vector2i(_viewer_id.index, _viewer_id.version.value))
+							.call_deferred();
+					// VoxelEngine::get_singleton().remove_viewer(_viewer_id);
+				}
 			}
 			break;
 
@@ -161,6 +175,25 @@ void VoxelViewer::_notification(int p_what) {
 		default:
 			break;
 	}
+}
+
+void VoxelViewer::unregister_deferred_callback(const ObjectID viewer_node_id, const Vector2i encoded_viewer_id) {
+	Object *obj = ObjectDB::get_instance(viewer_node_id);
+	VoxelViewer *viewer = Object::cast_to<VoxelViewer>(obj);
+	if (viewer != nullptr) {
+		viewer->_pending_deferred_unregistration = false;
+		if (viewer->is_inside_tree()) {
+			// The viewer is still in tree, so we can assume it was only reparented, so don't unregister it
+			// TODO When multi-world becomes supported, we have to re-check if the viewer has changed world...
+			return;
+		}
+	}
+	// The node got removed and not added back, or was destroyed
+	ViewerID viewer_id;
+	viewer_id.index = encoded_viewer_id.x;
+	viewer_id.version.value = encoded_viewer_id.y;
+	ZN_ASSERT_RETURN(VoxelEngine::get_singleton().viewer_exists(viewer_id));
+	VoxelEngine::get_singleton().remove_viewer(viewer_id);
 }
 
 bool VoxelViewer::is_active() const {